+++ R/aaa.R
@@ -1,0 +1,9 @@
+the <- new.env(parent = emptyenv())
+the$eyetracker_properties <- 
+  list(sample_frequency = NULL, 
+       viewing_distance_cm = 60,
+       screen_width_cm = 53,
+       screen_height_cm = 30,
+       screen_width_pixels = 1920,
+       screen_height_pixels = 1080)
+