+++ R/plot_spatial.R
@@ -6,11 +6,10 @@
 #' @param raw_data data in standard raw data form (time, x, y, trial)
 #' @param fix_data data output from fixation function
 #' @param sac_data data output from saccade function
-#' @param AOIs A dataframe of areas of interest (AOIs), with one row per AOI (x, y, width_radius, height). If using circular AOIs, then the 3rd column is used for the radius and the height should be set to NA.
+#' @param AOIs A dataframe of areas of interest (AOIs), with one row per AOI (name, x, y, width_radius, height). If using circular AOIs, then the column width_radius is used for the radius and the height should be set to NA.
 #' @param pID_values specify particular values within 'pID' to plot data from certain participants
 #' @param trial_values specify particular values within 'trial' to plot data from certain trials
 #' @param bg_image The filepath of a PNG image to be added to the plot, for example to show a screenshot of the task.
-#' @param res resolution of the display to be shown, as a vector (xmin, xmax, ymin, ymax)
 #' @param flip_y reverse the y axis coordinates (useful if origin is top of the screen)
 #' @param show_fix_order label the fixations in the order they were made
 #' @param plot_header display the header title text which explains graphical features of the plot.
@@ -44,7 +43,6 @@ plot_spatial <- function(raw_data = NULL,
                          pID_values = NULL,
                          trial_values = NULL,
                          bg_image = NULL,
-                         res = c(0,1920,0,1080),
                          flip_y = FALSE,
                          show_fix_order = TRUE,
                          plot_header = FALSE) {
@@ -54,36 +52,37 @@ plot_spatial <- function(raw_data = NULL,
   final_g <- ggplot()
 
   # setting axes limits and reversing y
-
-  if (is.null(res)==FALSE) {
-    # creates breaks based on quarters. Might look messy with some resolutions
-    breaks_x = round(seq(res[1],res[2],(res[2]-res[1])/4),0)
-    breaks_y = round(seq(res[3],res[4],(res[4]-res[3])/4),0)
-  }
-
-  if (is.null(res)==FALSE && flip_y==FALSE) {
-    final_g <- final_g +
-      scale_x_continuous(limits = res[1:2],
-                         breaks = breaks_x) +
-      scale_y_continuous(limits = res[3:4],
+  res_x <- the$eyetracker_properties$screen_width_pixels
+  res_y <- the$eyetracker_properties$screen_height_pixels
+  
+  # creates breaks based on quarters. Might look messy with some resolutions
+  breaks_x <- round(seq(0,res_x,res_x/4),0)
+  breaks_y <- round(seq(0,res_y,res_y/4),0)
+  
+  final_g <- final_g +
+      scale_x_continuous(limits = c(0,res_x),
+                         breaks = breaks_x)
+  if (flip_y==TRUE) {
+    final_g <- 
+      final_g +
+      scale_y_reverse(limits = c(res_y,0),
+                      breaks = rev(breaks_y)) 
+  } else {
+    final_g <- 
+      final_g +
+      scale_y_continuous(limits = c(0,res_y),
                          breaks = breaks_y)
-  } else if (is.null(res)==FALSE && flip_y==TRUE) {
-    final_g <- final_g +
-      scale_x_continuous(limits = res[1:2],
-                         breaks = breaks_x) +
-      scale_y_reverse(limits = res[4:3],
-                      breaks = breaks_y)
   }
 
   # PLOT BACKGROUND IMAGE
-  if (is.null(bg_image)==FALSE) final_g <- add_BGimg(bg_image, res, final_g)
+  if (is.null(bg_image)==FALSE) final_g <- add_BGimg(bg_image, flip_y, final_g)
 
   # PLOT gridlines
 
   # major gridlines are just the breaks_*
   # minor are [0:34 + half the diff
-  minor_breaks_x <- breaks_x[0:4] + ((res[2]-res[1])/8)
-  minor_breaks_y <- breaks_y[0:4] + ((res[4]-res[3])/8)
+  minor_breaks_x <- breaks_x[0:4] + (res_x/8)
+  minor_breaks_y <- breaks_y[0:4] + (res_y/8)
 
   final_g <-
     final_g +
@@ -101,7 +100,14 @@ plot_spatial <- function(raw_data = NULL,
     raw_data <- .select_pID_values(raw_data, pID_values, allow_random = FALSE)
     raw_data <- .select_trial_values(raw_data, trial_values, allow_random = FALSE)
     
-    final_g <- add_raw(raw_data, final_g)
+    final_g <- 
+      final_g +
+      geom_point(data = raw_data,
+                 aes(x = x, y = y),
+                 shape = 16,
+                 size = 3,
+                 alpha = .5,
+                 na.rm = TRUE)
   }
 
   # PLOT FIXATION DATA
@@ -188,28 +194,7 @@ plot_spatial <- function(raw_data = NULL,
            subtitle = "Raw data shown as dots; Fixations shown as circles (fill = duration); \nFixation size reflects dispersion of raw data; \nAOIs shown as blue regions")
   }
 
-
-
+  
   return(final_g)
-
-
-}
-
-# function to add raw data
-add_raw <- function(dataIn, ggplot_in){
-
-  x <- dataIn$x
-  y <- dataIn$y
-
-  ggplot_in <-
-    ggplot_in +
-    geom_point(data = dataIn,
-               aes(x = x, y = y),
-               shape = 16,
-               size = 3,
-               alpha = .5,
-               na.rm = TRUE)
-
-  return(ggplot_in)
 }